refactor: 슬라이더/버튼/체크박스 선택 컴포넌트화
